+++ infomeasure/estimators/entropy/__init__.py
@@ -3,6 +3,7 @@
 from .discrete import DiscreteEntropyEstimator
 from .kernel import KernelEntropyEstimator
 from .kozachenko_leonenko import KozachenkoLeonenkoEntropyEstimator
+from .miller_madow import MillerMadowEntropyEstimator
 from .renyi import RenyiEntropyEstimator
 from .ordinal import OrdinalEntropyEstimator
 from .tsallis import TsallisEntropyEstimator
@@ -11,6 +12,7 @@ __all__ = [
     "DiscreteEntropyEstimator",
     "KernelEntropyEstimator",
     "KozachenkoLeonenkoEntropyEstimator",
+    "MillerMadowEntropyEstimator",
     "OrdinalEntropyEstimator",
     "RenyiEntropyEstimator",
     "TsallisEntropyEstimator",